+++ Sources/Navigator/EPUB/EPUBSpreadView.swift
@@ -30,6 +30,9 @@ protocol EPUBSpreadViewDelegate: AnyObject {
     /// Called when the pages visible in the spread changed.
     func spreadViewPagesDidChange(_ spreadView: EPUBSpreadView)
 
+    /// Called when the user scrolls through the content.
+    func spreadView(_ spreadView: EPUBSpreadView, didScrollIn direction: ScrollDirection)
+
     /// Called when the spread view needs to present a view controller.
     func spreadView(_ spreadView: EPUBSpreadView, present viewController: UIViewController)